+++ bundles/org.corpus_tools.hexatomic.core/src/main/java/org/corpus_tools/hexatomic/core/events/salt/LabelNotifierImpl.java
@@ -24,6 +24,7 @@ package org.corpus_tools.hexatomic.core.events.salt;
 import org.corpus_tools.hexatomic.core.Topics;
 import org.corpus_tools.hexatomic.core.undo.operations.LabelModifyOperation;
 import org.corpus_tools.salt.graph.Label;
+import org.corpus_tools.salt.graph.LabelableElement;
 import org.corpus_tools.salt.graph.impl.LabelImpl;
 import org.eclipse.e4.core.services.events.IEventBroker;
 
@@ -88,6 +89,19 @@ public class LabelNotifierImpl extends LabelImpl
     modified(() -> super.setValue(value));
   }
 
+  @Override
+  public LabelableElement getContainer() {
+    LabelableElement container = super.getContainer();
+    if (container instanceof NotifyingElement<?>) {
+      NotifyingElement<?> notifyingContainer = (NotifyingElement<?>) container;
+      Object typedContainerDelegation = notifyingContainer.getTypedDelegation();
+      if (typedContainerDelegation instanceof LabelableElement) {
+        return (LabelableElement) typedContainerDelegation;
+      }
+    }
+    return super.getContainer();
+  }
+
   protected void modified(GraphModificationAction action) {
     if (this.getContainer() != null) {
       LabelModifyOperation op =
@@ -98,5 +112,4 @@ public class LabelNotifierImpl extends LabelImpl
       action.apply();
     }
   }
-
 }