+++ bundles/org.corpus_tools.hexatomic.corpusedit/src/main/java/org/corpus_tools/hexatomic/corpusedit/CorpusStructureView.java
@@ -207,10 +207,7 @@ public class CorpusStructureView {
     txtFilter.addModifyListener((ModifyEvent e) -> treeViewer.refresh());
 
     Composite composite = new Composite(parent, SWT.NONE);
-    GridData gridComposite = new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1);
-    gridComposite.minimumWidth = 300;
-    gridComposite.minimumHeight = 200;
-    composite.setLayoutData(gridComposite);
+    composite.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1));
     composite.setLayout(new TreeColumnLayout());
 
     treeViewer = new TreeViewer(composite, SWT.BORDER);