+++ bundles/org.corpus_tools.hexatomic.grid/src/main/java/org/corpus_tools/hexatomic/grid/internal/configuration/BodyMenuConfiguration.java
@@ -21,13 +21,20 @@
 
 package org.corpus_tools.hexatomic.grid.internal.configuration;
 
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
 import org.corpus_tools.hexatomic.grid.GridEditor;
-import org.corpus_tools.hexatomic.grid.internal.GridHelper;
+import org.corpus_tools.hexatomic.grid.internal.actions.CreateSpanSelectionAction;
+import org.corpus_tools.hexatomic.grid.internal.commands.DisplayAnnotationRenameDialogOnCellsCommand;
+import org.corpus_tools.hexatomic.grid.style.StyleConfiguration;
 import org.eclipse.nebula.widgets.nattable.NatTable;
 import org.eclipse.nebula.widgets.nattable.config.AbstractUiBindingConfiguration;
 import org.eclipse.nebula.widgets.nattable.coordinate.PositionCoordinate;
 import org.eclipse.nebula.widgets.nattable.edit.action.DeleteSelectionAction;
 import org.eclipse.nebula.widgets.nattable.grid.GridRegion;
+import org.eclipse.nebula.widgets.nattable.layer.LabelStack;
 import org.eclipse.nebula.widgets.nattable.selection.SelectionLayer;
 import org.eclipse.nebula.widgets.nattable.ui.NatEventData;
 import org.eclipse.nebula.widgets.nattable.ui.binding.UiBindingRegistry;
@@ -55,6 +62,7 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
   private final SelectionLayer selectionLayer;
 
   private static final String CHANGE_CELL_ANNOTATION_NAME_ITEM = "CHNG_ANNO_NAME"; //$NON-NLS-1$
+  private static final String CREATE_SPAN_ITEM = "CREATE_SPAN_ITEM";
 
   /**
    * Constructor setting the table and selection layer fields, and creating the menu via
@@ -71,11 +79,17 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
   }
 
   private Menu createMenu() {
-    ValidSelectionState validSelectionState = new ValidSelectionState();
     PopupMenuBuilder builder = new PopupMenuBuilder(this.table);
     builder.withMenuItemProvider(DELETE_CELL_ITEM, new DeleteItemProvider());
+    ValidSelectionState validSelectionState = new ValidSelectionState();
     builder.withVisibleState(DELETE_CELL_ITEM, validSelectionState);
+    builder.withMenuItemProvider(CHANGE_CELL_ANNOTATION_NAME_ITEM,
+        new ChangeAnnotationNameItemProvider());
     builder.withVisibleState(CHANGE_CELL_ANNOTATION_NAME_ITEM, validSelectionState);
+    builder.withMenuItemProvider(CREATE_SPAN_ITEM, new CreateSpanItemProvider());
+    ValidSingleSpanColumnEmptySelectionState validSingleSpanColumnEmptySelectionState =
+        new ValidSingleSpanColumnEmptySelectionState();
+    builder.withVisibleState(CREATE_SPAN_ITEM, validSingleSpanColumnEmptySelectionState);
     return builder.build();
   }
 
@@ -92,7 +106,7 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
    * 
    * @author Stephan Druskat (mail@sdruskat.net)
    */
-  public class DeleteItemProvider implements IMenuItemProvider {
+  private class DeleteItemProvider implements IMenuItemProvider {
 
 
     @Override
@@ -106,6 +120,87 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
           new DeleteSelectionAction().run(natTable, null);
         }
       });
+    }
+
+  }
+
+  /**
+   * Provides a menu item for changing annotation names on selected cells.
+   * 
+   * @author Stephan Druskat {@literal <mail@sdruskat.net>}
+   */
+  private class ChangeAnnotationNameItemProvider implements IMenuItemProvider {
+
+    @Override
+    public void addMenuItem(NatTable natTable, Menu popupMenu) {
+      MenuItem item = new MenuItem(popupMenu, SWT.PUSH);
+      item.setText(GridEditor.CHANGE_ANNOTATION_NAME_POPUP_MENU_LABEL);
+      item.setEnabled(true);
+      item.addSelectionListener(new SelectionAdapter() {
+        @Override
+        public void widgetSelected(SelectionEvent event) {
+          natTable.doCommand(
+              new DisplayAnnotationRenameDialogOnCellsCommand(natTable, createCellMapByColumn()));
+        }
+
+        private Map<Integer, Set<Integer>> createCellMapByColumn() {
+          // Map the selected cells by column.
+          Map<Integer, Set<Integer>> cellMapByColumn = new HashMap<>();
+          for (PositionCoordinate cellCoordinate : getSelectedNonTokenCells()) {
+            Set<Integer> columnCells = cellMapByColumn.get(cellCoordinate.getColumnPosition());
+            if (columnCells == null) {
+              columnCells = new HashSet<>();
+            }
+            columnCells.add(cellCoordinate.getRowPosition());
+            cellMapByColumn.put(cellCoordinate.getColumnPosition(), columnCells);
+          }
+          return cellMapByColumn;
+        }
+      });
+    }
+
+    private Set<PositionCoordinate> getSelectedNonTokenCells() {
+      Set<PositionCoordinate> selectedNonTokenCells = new HashSet<>();
+      PositionCoordinate[] selectedCellCoordinates = selectionLayer.getSelectedCellPositions();
+      for (PositionCoordinate cellPosition : selectedCellCoordinates) {
+        if (!isTokenCell(cellPosition)) {
+          selectedNonTokenCells.add(cellPosition);
+        }
+      }
+      return selectedNonTokenCells;
+    }
+
+  }
+
+  /**
+   * Provides a menu item for creating spans on selected cells.
+   * 
+   * @author Stephan Druskat {@literal <mail@sdruskat.net>}
+   */
+  private class CreateSpanItemProvider implements IMenuItemProvider {
+
+    @Override
+    public void addMenuItem(NatTable natTable, Menu popupMenu) {
+      MenuItem item = new MenuItem(popupMenu, SWT.PUSH);
+      item.setText(GridEditor.CREATE_SPAN_POPUP_MENU_LABEL);
+      item.setEnabled(true);
+      item.addSelectionListener(new SelectionAdapter() {
+        @Override
+        public void widgetSelected(SelectionEvent event) {
+          new CreateSpanSelectionAction(getSelectedNonTokenCells()).run(natTable);
+        }
+      });
+    }
+
+    private Set<PositionCoordinate> getSelectedNonTokenCells() {
+      Set<PositionCoordinate> selectedNonTokenCells = new HashSet<>();
+      PositionCoordinate[] selectedCellCoordinates = selectionLayer.getSelectedCellPositions();
+      for (PositionCoordinate cellPosition : selectedCellCoordinates) {
+        if (!isTokenCell(cellPosition)) {
+          selectedNonTokenCells.add(cellPosition);
+        }
+      }
+      return selectedNonTokenCells;
     }
 
   }
@@ -128,11 +223,9 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
         return false;
       } else {
         PositionCoordinate[] selectedCellCoordinates = selectionLayer.getSelectedCellPositions();
-
         for (PositionCoordinate coord : selectedCellCoordinates) {
           // Check whether the column at the position is the token column
-          if (GridHelper.isTokenColumnAtPosition(natEventData.getNatTable(),
-              coord.getColumnPosition(), false)) {
+          if (isTokenCell(coord)) {
             return false;
           }
         }
@@ -141,4 +234,55 @@ public class BodyMenuConfiguration extends AbstractUiBindingConfiguration {
     }
   }
 
+  /**
+   * A menu item state based on valid selection of cells.
+   * 
+   * <p>
+   * {@link #isActive(NatEventData)} returns <code>true</code> only when all selected cells are
+   * within a single span column, and all cells are empty.
+   * </p>
+   * 
+   * @author Stephan Druskat (mail@sdruskat.net)
+   */
+  private class ValidSingleSpanColumnEmptySelectionState implements IMenuItemState {
+
+    @Override
+    public boolean isActive(NatEventData natEventData) {
+      if (selectionLayer.getSelectedCells().isEmpty()) {
+        return false;
+      } else {
+        PositionCoordinate[] selectedCellCoordinates = selectionLayer.getSelectedCellPositions();
+
+        int singleColumnPosition = -1;
+        for (PositionCoordinate coord : selectedCellCoordinates) {
+          int columnPosition = coord.getColumnPosition();
+          int rowPosition = coord.getRowPosition();
+          // Check for each coordinate pair whether it has the same column position as the first
+          // pair (otherwise the cell is in a different column).
+          if (singleColumnPosition == -1) {
+            singleColumnPosition = columnPosition;
+          } else if (columnPosition != singleColumnPosition) {
+            return false;
+          }
+          if (selectionLayer.getDataValueByPosition(columnPosition, rowPosition) != null) {
+            return false;
+          }
+        }
+        // At this point, singleColumnPosition should be set
+        // Return whether the single column is a span column
+        return isSpanColumn(singleColumnPosition);
+      }
+    }
+
+    private boolean isSpanColumn(int singleColumnPosition) {
+      LabelStack configLabels = selectionLayer.getConfigLabelsByPosition(singleColumnPosition, 0);
+      return configLabels.getLabels().contains(StyleConfiguration.SPAN_ANNOTATION_CELL_STYLE);
+    }
+  }
+
+  private boolean isTokenCell(PositionCoordinate cellPosition) {
+    LabelStack configLabels = selectionLayer
+        .getConfigLabelsByPosition(cellPosition.getColumnPosition(), cellPosition.getRowPosition());
+    return configLabels.getLabels().contains(StyleConfiguration.TOKEN_TEXT_CELL_STYLE);
+  }
 }