+++ bundles/org.corpus_tools.hexatomic.grid/src/main/java/org/corpus_tools/hexatomic/grid/internal/data/GraphDataProvider.java
@@ -26,14 +26,18 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
 import java.util.TreeMap;
+import java.util.stream.Collectors;
 import javax.inject.Inject;
+import org.apache.commons.lang3.tuple.Pair;
 import org.corpus_tools.hexatomic.core.ProjectManager;
 import org.corpus_tools.hexatomic.core.errors.ErrorService;
 import org.corpus_tools.hexatomic.core.errors.HexatomicRuntimeException;
 import org.corpus_tools.hexatomic.grid.internal.data.Column.ColumnType;
+import org.corpus_tools.hexatomic.grid.internal.ui.UnrenamedAnnotationsDialog;
 import org.corpus_tools.salt.common.SDocumentGraph;
 import org.corpus_tools.salt.common.SSpan;
 import org.corpus_tools.salt.common.SSpanningRelation;
@@ -43,15 +47,17 @@ import org.corpus_tools.salt.common.STextualRelation;
 import org.corpus_tools.salt.common.SToken;
 import org.corpus_tools.salt.core.SAnnotation;
 import org.corpus_tools.salt.core.SRelation;
+import org.corpus_tools.salt.exceptions.SaltInsertionException;
+import org.corpus_tools.salt.util.SaltUtil;
 import org.eclipse.e4.core.di.annotations.Creatable;
 import org.eclipse.nebula.widgets.nattable.NatTable;
+import org.eclipse.nebula.widgets.nattable.coordinate.PositionCoordinate;
 import org.eclipse.nebula.widgets.nattable.data.IDataProvider;
 
 /**
  * Enables the use of an {@link SDocumentGraph} as a data source for the {@link NatTable}.
  * 
  * @author Stephan Druskat (mail@sdruskat.net)
- *
  */
 @Creatable
 public class GraphDataProvider implements IDataProvider {
@@ -387,6 +393,98 @@ public class GraphDataProvider implements IDataProvider {
     }
   }
 
+  /**
+   * Renames annotation on a set of given cells in bulk. The cells are identified by their
+   * {@link PositionCoordinate}s, which are provided as a map of cell positions to sets of row
+   * positions.
+   * 
+   * @param cellMapByColumn a map of cells from column positions to sets of row positions.
+   * @param newQName the new qualified annotation name for the annotations to be renamed
+   */
+  public void bulkRenameAnnotations(Map<Integer, Set<Integer>> cellMapByColumn, String newQName) {
+    Set<SStructuredNode> touchedNodes = new HashSet<>();
+    Set<SStructuredNode> unchangedNodes = new HashSet<>();
+    Pair<String, String> namespaceNamePair = SaltUtil.splitQName(newQName);
+    final String namespace = namespaceNamePair.getLeft();
+    final String name = namespaceNamePair.getRight();
+    // Run the rename for all cells by column
+    for (Entry<Integer, Set<Integer>> columnCoordinates : cellMapByColumn.entrySet()) {
+      Integer columnPosition = columnCoordinates.getKey();
+      Column column = getColumns().get(columnPosition);
+      String currentQName = column.getColumnValue();
+      if (currentQName.equals(newQName)) {
+        // Simply ignore this column, as the name is the same
+        log.debug("Ignoring rename operation on column {}, "
+            + "as the current and new qualified annotation names are the same: "
+            + "[current: {}]..[new: {}].", columnPosition, currentQName, newQName);
+        continue;
+      }
+      for (Integer rowPosition : columnCoordinates.getValue()) {
+        SStructuredNode node = column.getDataObject(rowPosition);
+        if (node != null && !touchedNodes.contains(node)) {
+          // Only proceed if the node is not null and if the node hasn't been touched yet.
+          SAnnotation currentAnnotation = node.getAnnotation(currentQName);
+          // Check if target annotation already exists
+          if (node.getAnnotation(newQName) != null) {
+            log.debug(
+                "The following node already has an annotation with the qualified name '{}'. "
+                    + "Ignoring it to avoid throwing {}:\n{}",
+                newQName, SaltInsertionException.class.getSimpleName(), node);
+            unchangedNodes.add(node);
+            continue;
+          }
+          Object annotationValue = currentAnnotation.getValue();
+          node.removeLabel(currentQName);
+          node.createAnnotation(namespace, name, annotationValue);
+          // Remember that the node has already been touched.
+          touchedNodes.add(node);
+        }
+      }
+    }
+    if (!unchangedNodes.isEmpty()) {
+      UnrenamedAnnotationsDialog.open(namespace, name, unchangedNodes);
+    }
+    projectManager.addCheckpoint();
+
+  }
+
+  /**
+   * Creates a new {@link SSpan} over the tokens in the first {@link Column} in {@link #columns},
+   * that are defined by the passed {@link PositionCoordinate#getColumnPosition()}s. Then, adds an
+   * {@link SAnnotation} to the new span with an empty value.
+   * 
+   * @param selectedCoordinates a set of the {@link PositionCoordinate}s of the currently selected
+   *        cells
+   */
+  public void createEmptyAnnotationSpan(Set<PositionCoordinate> selectedCoordinates) {
+    // Get tokens
+    List<Integer> selectedRows = selectedCoordinates.parallelStream()
+        .map(PositionCoordinate::getRowPosition).collect(Collectors.toList());
+    List<SStructuredNode> potentialTokens = selectedRows.parallelStream()
+        .map(i -> getColumns().get(0).getDataObject(i)).collect(Collectors.toList());
+    // Check that all potentialTokens are in fact tokens
+    List<SToken> tokens = potentialTokens.parallelStream().map(n -> {
+      if (n instanceof SToken) {
+        return (SToken) n;
+      } else {
+        throw new HexatomicRuntimeException(
+            "Expected an object of type " + SToken.class.getSimpleName()
+                + " in the first column of the grid, but found a " + n.getClass().getSimpleName());
+      }
+    }).collect(Collectors.toList());
+    SSpan span = graph.createSpan(tokens);
+    log.debug("Created new span {}.", span);
+    int columnIndex = selectedCoordinates.iterator().next().getColumnPosition();
+    Column column = getColumns().get(columnIndex);
+
+    for (int rowIndex : selectedRows) {
+      column.setRow(rowIndex, span);
+    }
+    createAnnotation(null, columnIndex, span);
+    log.debug("Annotated new span with empty value.");
+  }
+
+
   @Override
   public int getColumnCount() {
     return columns.size();
@@ -443,5 +541,4 @@ public class GraphDataProvider implements IDataProvider {
       return splitColumnValue[0];
     }
   }
-
 }